requirements.txt
alembic==1.4.3
astm==0.5.0
bcrypt==3.1.7
Bootstrap-Flask==1.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cssmin==0.2.0
dnspython==2.0.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RBAC==0.2.1
flask-redis==0.4.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lesscpy==0.14.0
Mako==1.1.3
MarkupSafe==1.1.1
netifaces==0.10.9
passlib==1.7.2
ply==3.11
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
visitor==0.1.3
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.1
